Improve compound fields documentation.
Since the `:==` is not yet supported for compound fields, I think it's
better to give an example with one of the _like_ operators.

The cheatsheet section on compound fields was a bit confusing, so I
tried to clean it up. I improved the layout and the wording of how the
operators work with compound fields.